osh_scale accepts floating-point targets

+++ src/osh_scale.cpp
@@ -11,7 +11,7 @@ int main(int argc, char** argv) {
     return -1;
   }
   auto path_in = argv[1];
-  auto target_nelems = atoi(argv[2]);
+  auto target_nelems = atof(argv[2]);
   auto path_out = argv[3];
   Omega_h::Mesh mesh(&lib);
   Omega_h::binary::read(path_in, lib.world(), &mesh);
